XIVY-18122 fix: new resource in context menu of cms entry
The URI of the item right clicked in the project explorer is used to determine the namespace for new resources created via the context menu.
The CMS entry has a URI pointing to the cms directory of the project.
Resolving the namespace failed if the cms directory did not exist.